Version 2.0 of VLM with Python 3

+++ pythonVLM/VLM_driver.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 # -*- coding: utf8 -*-
 # Copyright 2019 Université de Liège
 # 
@@ -22,7 +23,7 @@ class VLMDriver(object):
         self.data = CVLM.VLMData()
         CVLM.setup(infile, self.data)
         self.m = self.data.wing.nshed+2 # Assumes only wing
-        self.n = self.data.wing.nvert/self.m
+        self.n = int(self.data.wing.nvert/self.m)
         CVLM.geometry_setup(self.data)
         CVLM.cycleliftsurf(self.data)
         CVLM.memory_setup(self.data)
